LayoutTests/platform/qt/inspector/console/console-format-expected.txt

CONSOLE MESSAGE: line 12: Message format number %i, %d and %f
CONSOLE MESSAGE: line 13: Message %s for %s
CONSOLE MESSAGE: line 14: Object %o
CONSOLE MESSAGE: line 20: test,test2,,,test4
CONSOLE MESSAGE: line 21: %o
CONSOLE MESSAGE: line 22: %O
CONSOLE MESSAGE: line 54: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
CONSOLE MESSAGE: line 55: /^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
CONSOLE MESSAGE: line 54: /foo\\bar\sbaz/i
CONSOLE MESSAGE: line 55: /foo\\bar\sbaz/i
CONSOLE MESSAGE: line 54: test
CONSOLE MESSAGE: line 55: test
CONSOLE MESSAGE: line 54: test named "test"
CONSOLE MESSAGE: line 55: test named "test"
CONSOLE MESSAGE: line 54: Error
CONSOLE MESSAGE: line 55: Error
CONSOLE MESSAGE: line 54: [object HTMLBodyElement]
CONSOLE MESSAGE: line 55: [object HTMLBodyElement]
CONSOLE MESSAGE: line 54: function () { return 1; }
CONSOLE MESSAGE: line 55: function () { return 1; }
CONSOLE MESSAGE: line 54: function () {
        return 2;
    }
CONSOLE MESSAGE: line 55: function () {
        return 2;
    }
CONSOLE MESSAGE: line 54: 0.12
CONSOLE MESSAGE: line 55: 0.12
CONSOLE MESSAGE: line 54: http://webkit.org/
CONSOLE MESSAGE: line 55: http://webkit.org/
CONSOLE MESSAGE: line 54: null
CONSOLE MESSAGE: line 55: 
CONSOLE MESSAGE: line 54: undefined
CONSOLE MESSAGE: line 55: 
CONSOLE MESSAGE: line 54: [object Attr]
CONSOLE MESSAGE: line 55: [object Attr]
CONSOLE MESSAGE: line 54: [object Attr]
CONSOLE MESSAGE: line 55: [object Attr]
CONSOLE MESSAGE: line 54: [object Attr]
CONSOLE MESSAGE: line 55: [object Attr]
CONSOLE MESSAGE: line 54: [object Object]
CONSOLE MESSAGE: line 55: [object Object]
CONSOLE MESSAGE: line 54: NaN
CONSOLE MESSAGE: line 55: NaN
CONSOLE MESSAGE: line 54: Infinity
CONSOLE MESSAGE: line 55: Infinity
CONSOLE MESSAGE: line 54: -Infinity
CONSOLE MESSAGE: line 55: -Infinity
Tests that console logging dumps proper messages.

console-format.html:12Message format number 1, 2 and 3.5
console-format.html:13Message format for string
console-format.html:14Object Object
console-format.html:20["test", "test2", undefined, undefined, "test4"]
console-format.html:21["test", "test2", undefined, undefined, "test4"]
console-format.html:22Array[5]
console-format.html:54/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
console-format.html:55[/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i]
globals[0]
/^url\(\s*(?:(?:"(?:[^\\\"]|(?:\\[\da-f]{1,6}\s?|\.))*"|'(?:[^\\\']|(?:\\[\da-f]{1,6}\s?|\.))*')|(?:[!#$%&*-~\w]|(?:\\[\da-f]{1,6}\s?|\.))*)\s*\)/i
console-format.html:54/foo\\bar\sbaz/i
console-format.html:55[/foo\\bar\sbaz/i]
globals[1]
/foo\\bar\sbaz/i
console-format.html:54test
console-format.html:55["test"]
globals[2]
"test"
console-format.html:54test named "test"
console-format.html:55["test named "test""]
globals[3]
"test named "test""
console-format.html:54Error
console-format.html:55[Error]
globals[4]
Error
console-format.html:54<body onload="onload()">…</body>
console-format.html:55[<body onload="onload()">…</body>]
globals[5]
<body onload="onload()">…</body>
console-format.html:54function () { return 1; }
console-format.html:55[function () { return 1; }]
globals[6]
function () { return 1; }
console-format.html:54function () {
        return 2;
    }
console-format.html:55[function () {
        return 2;
    }]
globals[7]
function () {
        return 2;
    }
console-format.html:540.12
console-format.html:55[0.12]
globals[8]
0.12
console-format.html:54http://webkit.org/
console-format.html:55["http://webkit.org/"]
globals[9]
"http://webkit.org/"
console-format.html:54null
console-format.html:55[null]
globals[10]
null
console-format.html:54undefined
console-format.html:55[undefined]
globals[11]
undefined
console-format.html:54attr=""
console-format.html:55[attr=""]
globals[12]
attr=""
console-format.html:54attr="value"
console-format.html:55[attr="value"]
globals[13]
attr="value"
console-format.html:54id="x"
console-format.html:55[id="x"]
globals[14]
id="x"
console-format.html:54Object
console-format.html:55[Object]
globals[15]
Object
console-format.html:54NaN
console-format.html:55[NaN]
globals[16]
NaN
console-format.html:54Infinity
console-format.html:55[Infinity]
globals[17]
Infinity
console-format.html:54-Infinity
console-format.html:55[-Infinity]
globals[18]
-Infinity
